requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
PyJWT==2.4.0
SQLAlchemy==1.4.39
bcrypt==3.2.2